Refactor checkmark modal
