+++ src/components/chats/ChatItem/ChatItem.tsx
@@ -1,16 +1,12 @@
 import AddressAvatar from '@/components/AddressAvatar'
 import ProfilePreviewModalWrapper from '@/components/ProfilePreviewModalWrapper'
 import { isOptimisticId } from '@/services/datahub/posts/utils'
-import { useSendEvent } from '@/stores/analytics'
 import { useMessageData } from '@/stores/message'
 import { cx } from '@/utils/class-names'
 import { PostData } from '@subsocial/api/types'
-import { ComponentProps, SyntheticEvent, useReducer } from 'react'
+import { ComponentProps } from 'react'
 import ChatItemMenus from './ChatItemMenus'
 import ChatItemWithExtension from './ChatItemWithExtension'
-import CheckMarkExplanationModal, {
-  CheckMarkModalVariant,
-} from './CheckMarkExplanationModal'
 import Embed, { useCanRenderEmbed } from './Embed'
 import DefaultChatItem from './variants/DefaultChatItem'
 import EmojiChatItem, {
@@ -25,20 +21,6 @@ export type ChatItemProps = Omit<ComponentProps<'div'>, 'children'> & {
   enableChatMenu?: boolean
   chatId: string
   hubId: string
-}
-
-type CheckMarkModalReducerState = {
-  isOpen: boolean
-  variant: CheckMarkModalVariant | ''
-}
-const checkMarkModalReducer = (
-  state: CheckMarkModalReducerState,
-  action: CheckMarkModalVariant | ''
-): CheckMarkModalReducerState => {
-  if (action === '') {
-    return { ...state, isOpen: false }
-  }
-  return { isOpen: true, variant: action }
 }
 
 export default function ChatItem({
@@ -58,13 +40,6 @@ export default function ChatItem({
   const { createdAtBlock, ownerId, contentId } = message.struct
   const { body, extensions, link } = message.content || {}
 
-  const sendEvent = useSendEvent()
-
-  const [checkMarkModalState, dispatch] = useReducer(checkMarkModalReducer, {
-    isOpen: false,
-    variant: '',
-  })
-
   const setMessageAsReply = (messageId: string) => {
     if (isOptimisticId(messageId)) return
     setReplyTo(messageId)
@@ -73,15 +48,6 @@ export default function ChatItem({
   const canRenderEmbed = useCanRenderEmbed(link ?? '')
 
   if (!body && (!extensions || extensions.length === 0)) return null
-
-  const onCheckMarkClick = (e: SyntheticEvent) => {
-    e.stopPropagation()
-    const checkMarkType: CheckMarkModalVariant = isSent
-      ? 'recorded'
-      : 'recording'
-    sendEvent('click check_mark_button', { type: checkMarkType })
-    dispatch(checkMarkType)
-  }
 
   const isEmojiOnly = shouldRenderEmojiChatItem(body ?? '')
   const ChatItemContentVariant = isEmojiOnly ? EmojiChatItem : DefaultChatItem
@@ -128,7 +94,6 @@ export default function ChatItem({
               >
                 {extensions && extensions.length > 0 ? (
                   <ChatItemWithExtension
-                    onCheckMarkClick={onCheckMarkClick}
                     scrollToMessage={scrollToMessage}
                     message={message}
                     isMyMessage={isMyMessage}
@@ -139,7 +104,6 @@ export default function ChatItem({
                   <ChatItemContentVariant
                     message={message}
                     isMyMessage={isMyMessage}
-                    onCheckMarkClick={onCheckMarkClick}
                     scrollToMessage={scrollToMessage}
                     chatId={chatId}
                     hubId={hubId}
@@ -149,13 +113,6 @@ export default function ChatItem({
             )
           }}
         </ChatItemMenus>
-        <CheckMarkExplanationModal
-          isOpen={checkMarkModalState.isOpen}
-          variant={checkMarkModalState.variant || 'recording'}
-          closeModal={() => dispatch('')}
-          blockNumber={createdAtBlock}
-          cid={contentId}
-        />
       </div>
       {canRenderEmbed && (
         <div className={cx(isMyMessage ? 'flex justify-end' : 'flex')}>
