+++ src/components/extensions/donate/DonateMessagePreview.tsx
@@ -79,7 +79,6 @@ type DonateMessagePreviewProps = ExtensionChatItemProps
 
 export default function DonateMessagePreview({
   message,
-  onCheckMarkClick,
   scrollToMessage,
   chatId,
   hubId,
@@ -97,7 +96,6 @@ export default function DonateMessagePreview({
   return (
     <CommonChatItem
       message={message}
-      onCheckMarkClick={onCheckMarkClick}
       scrollToMessage={scrollToMessage}
       myMessageConfig={{ children: 'bottom', checkMark: 'outside' }}
       className={cx('relative flex flex-col overflow-hidden')}
